Update JiebaSegmenter.java for Solr highlighting
I've done some minor modification in the code under JiebaSegmenter.java, and the highlighting in Solr seems to be fine now. The <em> tag are now at the correct place.

Basically, I created another int called offset2 under process() method.
int offset2 = 0; 

And I modified the offset to offset2 for this part of the code under process() method. I have indicated in the comments for the changes that I made.